# src/Domivice.Users.Web/wwwroot/openapi-original.json
{
    "openapi": "3.0.2",
    "info": {
        "contact": {
            "name": "Domivice Development Team",
            "url": "https://development.domivice.com"
        },
        "description": "The users API",
        "title": "Users API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
    "paths": {
        "/v1/users/{userId}": {
            "delete": {
                "parameters": [
                    {
                        "description": "The user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Response when the user has been deleted"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not allowed to perform action"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "The resource was not found"
                    }
                },
                "summary": "Delete User",
                "tags": ["Users"]
            },
            "get": {
                "parameters": [
                    {
                        "description": "The user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        },
                        "description": "Response when a user object is returned"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "The resource was not found"
                    }
                },
                "summary": "Get User",
                "tags": ["Users"]
            },
            "patch": {
                "parameters": [
                    {
                        "description": "The user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        },
                        "description": "User update was successful "
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "There was an error with the request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not allowed to perform action"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "The resource was not found"
                    }
                },
                "summary": "Update User",
                "tags": ["Users"]
            }
        },
        "/v1/users": {
            "get": {
                "parameters": [
                    {
                        "description": "A search term to search users on firstName, lastName, email and phoneNumbers.",
                        "explode": true,
                        "in": "query",
                        "name": "search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "style": "form"
                    }, {
                        "description": "The status to filter users by status",
                        "explode": true,
                        "in": "query",
                        "name": "status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "style": "form"
                    }, {
                        "description":
                            "The page requested. Page parameter can only be a positive integer greater than 0.",
                        "explode": true,
                        "in": "query",
                        "name": "page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "style": "form"
                    }, {
                        "description": "The requested page size. Page size can only be between 1 and 100",
                        "explode": true,
                        "in": "query",
                        "name": "pageSize",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "style": "form"
                    }, {
                        "description":
                            "The result sorting. field1[:asc|desc][,field2[:asc|desc]] e.g. firstName:desc,email:asc",
                        "explode": true,
                        "in": "query",
                        "name": "sort",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "style": "form"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserList"
                                }
                            }
                        },
                        "description": "When a list of 0 or more users are returned"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    }
                },
                "summary": "List Users",
                "tags": ["Users"]
            }
        },
...
    },
    "components": {
...
        "schemas": {
            "Error": {
                "description": "",
                "properties": {
                    "detail": {
                        "description": "A human-readable explanation of the error",
                        "type": "string"
                    },
                    "instance": {
                        "description": "A URI that identifies the specific occurrence of the error",
                        "type": "string"
                    },
                    "status": {
                        "description": "The HTTP response code",
                        "type": "string"
                    },
                    "title": {
                        "description": "A brief, human-readable message about the error",
                        "type": "string"
                    },
                    "type": {
                        "description": "A URI identifier that categorizes the error",
                        "type": "string"
                    },
                    "errors": {
                        "description": "The validation errors",
                        "items": {
                            "$ref": "#/components/schemas/KeyValuePair"
                        },
                        "type": "array"
                    },
                    "extensions": {
                        "description": "The error extensions",
                        "items": {
                            "$ref": "#/components/schemas/KeyValuePair"
                        },
                        "type": "array"
                    }
                },
                "required": ["detail", "instance", "title", "type"],
                "type": "object"
            },
            "KeyValuePair": {
                "description": "Carries key values pair",
                "type": "object"
            },
...
            "User": {
                "description": "Carries information about the user",
                "example": {
                    "firstName": "firstName",
                    "lastName": "lastName",
                    "phoneNumber": "phoneNumber",
                    "id": "id",
                    "email": "email",
                    "status": "status"
                },
                "properties": {
                    "id": {
                        "description": "The string guid representing the user id ",
                        "type": "string"
                    },
                    "email": {
                        "description": "The user email",
                        "type": "string"
                    },
                    "firstName": {
                        "description": "The user first name",
                        "type": "string"
                    },
                    "lastName": {
                        "description": "The user last name",
                        "type": "string"
                    },
                    "status": {
                        "description": "The user status",
                        "type": "string"
                    },
                    "phoneNumber": {
                        "description": "The user phone number",
                        "type": "string"
                    }
                },
                "required": ["email", "firstName", "id", "lastName", "phoneNumber", "status"],
                "type": "object"
            },
            "UserList": {
                "description": "Carries a paginated list of users",
                "example": {
                    "totalItemsCount": 6,
                    "pageCount": 0,
                    "previousPage": "previousPage",
                    "data": [
                        {
                            "firstName": "firstName",
                            "lastName": "lastName",
                            "phoneNumber": "phoneNumber",
                            "id": "id",
                            "email": "email",
                            "status": "status"
                        }, {
                            "firstName": "firstName",
                            "lastName": "lastName",
                            "phoneNumber": "phoneNumber",
                            "id": "id",
                            "email": "email",
                            "status": "status"
                        }
                    ],
                    "nextPage": "nextPage"
                },
                "properties": {
                    "data": {
                        "description": "The user list",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        },
                        "type": "array"
                    },
                    "nextPage": {
                        "description": "The URL to access the next page",
                        "type": "string"
                    },
                    "pageCount": {
                        "description": "The page count",
                        "type": "integer"
                    },
                    "previousPage": {
                        "description": "The URL to access the previous page",
                        "type": "string"
                    },
                    "totalItemsCount": {
                        "description": "The total items count",
                        "type": "integer"
                    }
                },
                "required": ["data", "pageCount", "totalItemsCount"],
                "type": "object"
            },
...
            "UserUpdate": {
                "description": "Carries information about the user",
                "example": {
                    "firstName": "firstName",
                    "lastName": "lastName",
                    "phoneNumber": "phoneNumber",
                    "phoneCountryCode": "phoneCountryCode"
                },
                "properties": {
                    "firstName": {
                        "description": "The user first name",
                        "type": "string"
                    },
                    "lastName": {
                        "description": "The user last name",
                        "type": "string"
                    },
                    "phoneNumber": {
                        "description": "The user phone number",
                        "type": "string"
                    },
                    "phoneCountryCode": {
                        "description": "The user phone country code",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}